+++ src/mainwindow.h
@@ -277,6 +277,8 @@ private:
 
     void handleTextNoteLinking();
 
+    bool downloadUrlToFile(QUrl url, QFile *file);
+
     void updateNoteTextFromDisk(Note note);
 
     void makeCurrentNoteFirstInNoteList();